pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paddy"
version = "0.1.0"
description = "Deep Learning models"
authors = [
    {name = "GooLey", email = "gooleywm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "tensorflow>=2.15",
    "numpy",
    "pyyaml",
    "pandas",
    "h5py",
    "natsort",
    "tqdm",
    "google-cloud-storage",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
paddy_train = "paddy.scripts.paddy_train:main"
paddy_data = "paddy.scripts.paddy_data:main"
extract_label = "paddy.helpers.extract_label:main" 
